Refactoring and Separating into some modules
- Separate into main, http_resolver, dns_resolver, config
- Remove simple-json dependency
