+++ README.md
@@ -38,5 +38,4 @@ The 3-Clause BSD License. See also LISENCE file.
 ### statically linked libraries:
 
 - [golang/go](https://github.com/golang/go/) ... [BSD 3-clause "New" or "Revised" License](https://github.com/golang/go/blob/master/LICENSE)
-- [bitly/go-simplejson](https://github.com/bitly/go-simplejson) ... [MIT License](https://github.com/bitly/go-simplejson/blob/master/LICENSE)
 - [miekg/dns](https://github.com/miekg/dns) ... [BSD 3-clause "New" or "Revised" License](https://github.com/miekg/dns/blob/master/LICENSE)
